major architectural changes to abstract logic to reusable hooks
