+++ app/(tabs)/builder/build-step-1.tsx
@@ -1,179 +1,6 @@
-import { useCurrentNPC } from '@/app/context/current-npc';
-import LoadingScreen from '@/components/LoadingScreen';
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
-import { getAges, getGenders, getPotencies } from '@/services/build-step-1';
-import { Age, Gender, Potency } from '@/types/db-schema';
-import { useSQLiteContext } from 'expo-sqlite';
-import { useEffect, useState } from 'react';
-import { ScrollView, StyleSheet, TouchableOpacity } from 'react-native';
+import BuildStep1 from "@/components/build-steps/build-step-1";
 
-// FIXME: state resets on page change. Will need global context
-// TODO: Abstract most of this logic to a component and just house the route here
 
-export default function BuildStep1() {
-  const db = useSQLiteContext();
-  const { currentNPC, updateCurrentNPC } = useCurrentNPC();
-
-  const [genders, setGenders] = useState<Gender[]>([]);
-  const [ages, setAges] = useState<Age[]>([]);
-  const [potencies, setPotencies] = useState<Potency[]>([]);
-
-  const [selectedGenderId, setSelectedGenderId] = useState<number>(currentNPC?.gender_id || 2);
-  const [selectedAgeId, setSelectedAgeId] = useState<number>(currentNPC?.age_id || 2);
-  const [selectedPotencyId, setSelectedPotencyId] = useState<number>(currentNPC?.potency_id || 3);
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-  
-  
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        setIsLoading(true);
-        
-        const [gendersResponse, agesResponse, potenciesResponse] = await Promise.all([
-          getGenders(db),
-          getAges(db),
-          getPotencies(db)
-        ]);
-
-        if (gendersResponse.data) {
-          setGenders(gendersResponse.data);
-        }
-        
-        if (agesResponse.data) {
-          setAges(agesResponse.data);
-        }
-        
-        if (potenciesResponse.data) {
-          setPotencies(potenciesResponse.data);
-        }
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    fetchData();
-  }, [db]);
-
-    const handleGenderSelect = async (genderId: number) => {
-    setSelectedGenderId(genderId);
-    await updateCurrentNPC({ gender_id: genderId });
-  };
-
-  const handleAgeSelect = async (ageId: number) => {
-    setSelectedAgeId(ageId);
-    await updateCurrentNPC({ age_id: ageId });
-  };
-
-  const handlePotencySelect = async (potencyId: number) => {
-    setSelectedPotencyId(potencyId);
-    await updateCurrentNPC({ potency_id: potencyId });
-  };
-
-  const renderSelectionButtons = <T extends { id: number; name: string }>(
-    options: T[],
-    selectedId: number,
-    onSelect: (id: number) => void,
-    horizontal: boolean = true
-  ) => (
-    <ThemedView style={horizontal ? styles.horizontalContainer : styles.verticalContainer}>
-      {options.map((option) => (
-        <TouchableOpacity
-          key={option.id}
-          style={[
-            styles.selectionButton,
-            selectedId === option.id && styles.selectedButton,
-            !horizontal && styles.fullWidthButton
-          ]}
-          onPress={() => onSelect(option.id)}
-        >
-          <ThemedText style={[
-            styles.buttonText,
-            selectedId === option.id && styles.selectedButtonText
-          ]}>
-            {option.name}
-          </ThemedText>
-        </TouchableOpacity>
-      ))}
-    </ThemedView>
-  );
-
-  if (isLoading) {
-    return (
-      <LoadingScreen />
-    );
-  }
-
-  return (
-    <ScrollView style={styles.container}>      
-      {/* Gender Section */}
-      <ThemedView style={styles.sectionContainer}>
-        <ThemedText style={styles.sectionText} type="subtitle">Gender</ThemedText>
-        {renderSelectionButtons(genders, selectedGenderId, handleGenderSelect)}
-      </ThemedView>
-
-      {/* Age Section */}
-      <ThemedView style={styles.sectionContainer}>
-        <ThemedText style={styles.sectionText} type="subtitle">Age</ThemedText>
-        {renderSelectionButtons(ages, selectedAgeId, handleAgeSelect)}
-      </ThemedView>
-
-      {/* Potency Section */}
-      <ThemedView style={styles.sectionContainer}>
-        <ThemedText style={styles.sectionText} type="subtitle">Potency</ThemedText>
-        {renderSelectionButtons(potencies, selectedPotencyId, handlePotencySelect, false)}
-      </ThemedView>
-    </ScrollView>
-  );
+export default function BuildStep1Route() {
+  return <BuildStep1 />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 16,
-  },
-  titleContainer: {
-    flexDirection: 'row',
-    gap: 8,
-    marginBottom: 16,
-  },
-  sectionContainer: {
-    marginVertical: 20,
-  },
-  horizontalContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginTop: 10,
-  },
-  verticalContainer: {
-    marginTop: 10,
-  },
-  selectionButton: {
-    backgroundColor: '#D3D3D3',
-    paddingVertical: 12,
-    paddingHorizontal: 20,
-    borderRadius: 8,
-    flex: 1,
-    marginHorizontal: 4,
-    alignItems: 'center',
-  },
-  fullWidthButton: {
-    marginVertical: 4,
-    marginHorizontal: 0,
-  },
-  selectedButton: {
-    backgroundColor: '#007AFF',
-  },
-  buttonText: {
-    fontSize: 16,
-    fontWeight: '500',
-  },
-  selectedButtonText: {
-    color: 'white',
-  },
-  sectionText: {
-    textAlign: 'center'
-  }
-});
